# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Lelang Ikan API",
        "description": "API untuk aplikasi e-lelang ikan realtime",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local API Base"
        }
    ],
    "paths": {
...
        "/media": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Upload media (image/video)",
                "operationId": "6c4c309050b66c61c31a8a5bd68ae05c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/seller/commodities": {
            "get": {
                "tags": [
                    "Seller - Commodities"
                ],
                "summary": "List commodities milik seller",
                "operationId": "d871ca63254f3b5cd4dd9a7f60e65e0f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (Laravel paginator)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page. Max is capped by server.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Seller - Commodities"
                ],
                "summary": "Create commodity (seller)",
                "operationId": "40ca83353defe3f56911bb4ffbc3de7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ikan Tuna"
                                    },
                                    "weight_kg": {
                                        "type": "number",
                                        "example": 12.5
                                    },
                                    "size_grade": {
                                        "type": "string",
                                        "example": "L"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "TPI Buleleng"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft"
                                    },
                                    "media": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "image"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "/storage/uploads/a.jpg"
                                                },
                                                "sort_order": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/seller/commodities/{id}": {
            "patch": {
                "tags": [
                    "Seller - Commodities"
                ],
                "summary": "Update commodity (seller, milik sendiri)",
                "operationId": "555d1399aaa06bc89397014dfe9eddc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/rating-summary": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get rating summary for a user (public)",
                "operationId": "02c3a398f5df9235b659fe5d12874a53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan Bearer token dari endpoint login/register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}